moving directories to components
