+++ src/app/features/features.module.ts
@@ -28,7 +28,7 @@ import { EditActivityFormComponent } from "./backoffice/pages/activities/edit-ac
 import { CoreModule } from '../core/core.module';
 import { HomeComponent } from "./public/pages/home/home.component";
 import { IndexComponent } from "./public/pages/home/index/index.component";
-import { UsComponent } from "./public/pages/us/us.component";
+import { UsComponent } from './public/pages/about/components/us/us.component';
 @NgModule({
   declarations: [
     ActivityFormComponent,
